OPENEUROPA-1531: Remove base media value object.
