+++ oe_theme.theme
@@ -14,7 +14,7 @@ use Drupal\Component\Utility\Random;
 use Drupal\oe_theme\ValueObject\FileValueObject;
 use Drupal\oe_theme\ValueObject\DateValueObject;
 use Drupal\oe_theme\ValueObject\GalleryItemValueObject;
-use Drupal\oe_theme\ValueObject\ImageMediaValueObject;
+use Drupal\oe_theme\ValueObject\ImageValueObject;
 
 /**
  * Implements hook_preprocess_breadcrumb().
@@ -435,7 +435,7 @@ function oe_theme_preprocess_pattern_gallery(array &$variables): void {
       $item['icon'] = 'ecl-icon--' . $item['icon'];
     }
 
-    $item['image'] = ImageMediaValueObject::fromArray($item['image']);
+    $item['image'] = ImageValueObject::fromArray($item['image']);
     $variables['rows'][ceil($count_item / 4)][] = GalleryItemValueObject::fromArray($item);
     $count_item++;
   }
