--- tests/Unit/ValueObject/MediaValueObjectTest.php
@@ -1,31 +1,0 @@
-<?php
-
-declare(strict_types = 1);
-
-namespace Drupal\Tests\oe_theme\Unit\Patterns;
-
-use Drupal\oe_theme\ValueObject\MediaValueObject;
-use Drupal\Tests\UnitTestCase;
-
-/**
- * Test media value object.
- */
-class MediaValueObjectTest extends UnitTestCase {
-
-  /**
-   * Test constructing a media value object from an array.
-   */
-  public function testFromArray() {
-    $data = [
-      'src' => 'http://example.com/test.pdf',
-      'name' => 'Test.pdf',
-    ];
-
-    /** @var \Drupal\oe_theme\ValueObject\MediaValueObject $object */
-    $object = MediaValueObject::fromArray($data);
-
-    $this->assertEquals($data['src'], $object->getSource());
-    $this->assertEquals($data['name'], $object->getName());
-  }
-
-}
